pyproject.toml
[project]
name = "dio-blog"
version = "0.1.0"
description = "Projeto FastAPI para Luizalabs\""
authors = [
    {name = "guilhermesilvacorreia",email = "guilhermecorreiasilva29@gmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "databases[aiosqlite] (>=0.9.0,<0.10.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
